Do not require patch versions unless it is relay necessary
Even Ubuntu 20.04 LTS has php 7.4.3 at the moment, so this composer json blocks the installing process (or you have to use --ignore-platform-reqs). As patch versions should not bring new features but only solve bugs and security things, it is unlikely that the patch version is important. (See https://semver.org)